+++ ft_itoa.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/10 13:42:13 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/09 12:34:19 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/10 11:43:34 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_itoa(long value)
+char	*ft_itoa(long long value)
 {
 	return (ft_itoa_base(value, 10));
 }
