+++ ft_min.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/19 12:40:13 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/09 12:36:59 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/10 11:44:00 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_min(long n1, long n2)
+long long	ft_min(long long n1, long long n2)
 {
 	if (n1 < n2)
 		return (n1);
