+++ ft_numlen.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/03 14:52:39 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/09 12:31:44 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/10 11:45:07 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_numlen(long value)
+int		ft_numlen(long long value)
 {
 	return (ft_numlen_base(value, 10));
 }
