+++ ft_putnbr.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/08 15:28:35 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/09 12:33:15 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/10 11:45:49 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr(long n)
+void	ft_putnbr(long long n)
 {
 	ft_putnbr_fd_base(n, 1, 10);
 }
