+++ ft_putnbr_fd.c
@@ -6,13 +6,13 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/09 12:12:03 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/09 12:33:12 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/10 11:45:41 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr_fd(long n, int fd)
+void	ft_putnbr_fd(long long n, int fd)
 {
 	ft_putnbr_fd_base(n, fd, 10);
 }
